requirements.txt
Django~=5.1.2
requests~=2.32.3
python-dotenv~=1.0.1
python-decouple~=3.8
assemblyai
google-generativeai
